chore: migrate /rechercher to app folder
