+++ app/(header-search)/error.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+import { SearchErrorExplanations } from '#components/error-explanations';
+import {
+  NextAppError,
+  useLogFatalErrorAppClient,
+} from 'hooks/use-log-fatal-error-app-client';
+
+export default function ErrorPage({ error }: { error: NextAppError }) {
+  useLogFatalErrorAppClient(error);
+
+  return <SearchErrorExplanations />;
+}
