+++ components/dirigeants-section/rne-dirigeants.tsx
@@ -15,23 +15,7 @@ import {
   IPersonneMorale,
 } from '#models/immatriculation';
 import { formatDatePartial, formatIntFr } from '#utils/helpers';
-
-/**
- * Weird bug happennig here. Webpack build fail when this function is in model/dirigeants.ts
- * @param toBeDetermined
- * @returns
- */
-export const isPersonneMorale = (
-  toBeDetermined: IEtatCivil | IPersonneMorale
-): toBeDetermined is IPersonneMorale => {
-  if (
-    (toBeDetermined as IPersonneMorale).siren ||
-    (toBeDetermined as IPersonneMorale).denomination
-  ) {
-    return true;
-  }
-  return false;
-};
+import { isPersonneMorale } from './is-personne-morale';
 
 type IProps = {
   immatriculationRNE:
